api/requirements.txt
fastapi
uvicorn
pandas
mlflow
databricks-sdk
pydantic
xgboost
scikit-learn
python-dotenv